refactor: code clean up, prep for review and release
