+++ packages/cli/src/rpc/navie/thread/handlers/query.ts
@@ -1,16 +1,15 @@
 import { NavieRpc } from '@appland/rpc';
-import { RpcHandler } from '../../rpc';
-import { ThreadIndex } from '.';
+import { RpcHandler } from '../../../rpc';
+import { ThreadIndexService } from '../../services/threadIndexService';
 
-export function navieThreadQueryHandler(): RpcHandler<
-  NavieRpc.V1.Thread.Query.Params,
-  NavieRpc.V1.Thread.Query.Response
-> {
+export function navieThreadQueryHandler(
+  threadIndexService: ThreadIndexService
+): RpcHandler<NavieRpc.V1.Thread.Query.Params, NavieRpc.V1.Thread.Query.Response> {
   return {
     name: NavieRpc.V1.Thread.Query.Method,
     handler: ({ threadId, maxCreatedAt, orderBy, limit, offset, projectDirectories }) =>
       // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      ThreadIndex.getInstance().query({
+      threadIndexService.query({
         uuid: threadId,
         maxCreatedAt,
         orderBy,
