+++ packages/cli/src/rpc/navie/thread/handlers/sendMessage.ts
@@ -1,19 +1,18 @@
 import { NavieRpc } from '@appland/rpc';
-import { RpcHandler } from '../../rpc';
-import { getThread } from '.';
+import { RpcHandler } from '../../../rpc';
+import ThreadService from '../../services/threadService';
 
-export function navieThreadSendMessageHandler(): RpcHandler<
-  NavieRpc.V1.Thread.SendMessage.Params,
-  NavieRpc.V1.Thread.SendMessage.Response
-> {
+export function navieThreadSendMessageHandler(
+  threadService: ThreadService
+): RpcHandler<NavieRpc.V1.Thread.SendMessage.Params, NavieRpc.V1.Thread.SendMessage.Response> {
   return {
     name: NavieRpc.V1.Thread.SendMessage.Method,
     async handler({
       threadId,
       content,
-      codeSelection,
+      userContext,
     }): Promise<NavieRpc.V1.Thread.SendMessage.Response> {
-      const thread = await getThread(threadId);
+      const thread = await threadService.getThread(threadId);
       if (!thread) {
         const errorMessage = `Thread ${threadId} not found`;
         console.warn(errorMessage);
@@ -21,7 +20,7 @@ export function navieThreadSendMessageHandler(): RpcHandler<
       }
 
       try {
-        await thread.sendMessage(content, codeSelection);
+        await thread.sendMessage(content, userContext);
         return { ok: true };
       } catch (err) {
         return { ok: false, error: err };
