+++ packages/cli/src/rpc/navie/thread/index.ts
@@ -1,103 +1,38 @@
-import { ContextV2, Help, Navie, ProjectInfo, UserContext } from '@appland/navie';
+import { Navie, UserContext } from '@appland/navie';
 import { dirname, join } from 'path';
 import { EventEmitter } from 'stream';
 import { randomUUID } from 'crypto';
 import { ConversationThread } from '@appland/client';
 import { getSuggestions } from '../suggest';
-import { NavieRpc } from '@appland/rpc';
 import { homedir } from 'os';
-import { mkdir, readFile, writeFile } from 'fs/promises';
-import sqlite3 from 'better-sqlite3';
+import { mkdir, writeFile } from 'fs/promises';
+
 import NavieService from '../services/navieService';
-import configuration from '../../configuration';
+import { NavieEvent, NavieMessageEvent, PinnedItem, TimestampNavieEvent } from './events';
+import { ThreadIndexService } from '../services/threadIndexService';
+import { container } from 'tsyringe';
+import { NavieRpc } from '@appland/rpc';
+import handleReview from '../../explain/review';
 
-type NavieThreadInitEvent = {
-  type: 'thread-init';
-  conversationThread: ConversationThread;
-};
-
-type NavieTokenMetadataEvent = {
-  type: 'token-metadata';
-  codeBlockId: string;
-  metadata: Record<string, unknown>;
-};
-
-type NavieTokenEvent = {
-  type: 'token';
-  messageId: string;
-  token: string;
-  codeBlockId?: string;
-};
-
-type NavieMessageEvent = {
-  type: 'message';
-  role: 'system' | 'assistant' | 'user';
-  messageId: string;
-  content: string;
-};
-
-type NaviePromptSuggestionsEvent = {
-  type: 'prompt-suggestions';
-  suggestions: NavieRpc.V1.Suggest.Response;
-  messageId: string;
-};
-
-type NavieMessageCompleteEvent = {
-  type: 'message-complete';
-  messageId: string;
-};
-
-type NaviePinItemEvent = PinnedItem & {
-  type: 'pin-item';
-};
-
-type NavieUnpinItemEvent = PinnedItem & {
-  type: 'unpin-item';
-};
-
-type NavieErrorEvent = {
-  type: 'error';
-  error: unknown;
-};
-
-type NavieBeginContextSearchEvent = {
-  type: 'begin-context-search';
-  contextType: 'help' | 'project-info' | 'context';
-  id: string;
-  request?: Help.HelpRequest | ProjectInfo.ProjectInfoRequest | ContextV2.ContextRequest;
-};
-
-type NavieCompleteContextSearchEvent = {
-  type: 'complete-context-search';
-  id: string;
-  result?: Help.HelpResponse | ProjectInfo.ProjectInfoResponse | ContextV2.ContextResponse;
-};
-
-type Timestamp = {
-  time: number;
-};
-
-type NavieEvent =
-  | NavieBeginContextSearchEvent
-  | NavieCompleteContextSearchEvent
-  | NavieErrorEvent
-  | NavieMessageEvent
-  | NavieMessageCompleteEvent
-  | NaviePinItemEvent
-  | NavieUnpinItemEvent
-  | NaviePromptSuggestionsEvent
-  | NavieThreadInitEvent
-  | NavieTokenEvent
-  | NavieTokenMetadataEvent;
-
-type TimestampNavieEvent = Timestamp & NavieEvent;
-
-type PinnedItem = {
-  uri?: string;
-  handle?: number;
-};
-
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 type EventListener = (...args: any[]) => void;
+
+/**
+ * Converts the simplified context format back into the format expected by Navie.
+ * Why two formats? Because the difference between a `code-snippet` and a `code-selection` is too
+ * ambiguous, and `file` too specific. Ideally I'd like to migrate to use this static/dynamic format
+ * everywhere, but it works for now. - DB
+ */
+function convertContext(
+  context?: NavieRpc.V1.UserContext.ContextItem[]
+): undefined | UserContext.Context {
+  return context?.map((item) => {
+    if (item.type === 'static') {
+      return { type: 'code-snippet', content: item.content, location: item.id };
+    }
+    return { type: 'file', location: item.uri };
+  });
+}
 
 export class Thread {
   private eventEmitter = new EventEmitter();
@@ -109,7 +44,16 @@ export class Thread {
   private lastTokenBeganCodeBlock = false;
   private static readonly HISTORY_DIRECTORY = join(homedir(), '.appmap', 'navie', 'history');
 
-  constructor(public readonly conversationThread: ConversationThread) {}
+  constructor(
+    public readonly conversationThread: ConversationThread,
+    private readonly navieService: NavieService,
+    eventLog?: TimestampNavieEvent[]
+  ) {
+    if (eventLog?.length) {
+      this.log = eventLog;
+      this.lastEventWritten = this.log.length;
+    }
+  }
 
   initialize() {
     this.logEvent({ type: 'thread-init', conversationThread: this.conversationThread });
@@ -123,42 +67,10 @@ export class Thread {
 
   private async emitSuggestions(messageId: string) {
     const suggestions = await getSuggestions(
-      NavieService.getNavieProvider(),
+      this.navieService.getNavieProvider(),
       this.conversationThread.id
     );
     this.logEvent({ type: 'prompt-suggestions', suggestions, messageId });
-  }
-
-  static async load(threadId: string): Promise<Thread> {
-    const historyFilePath = Thread.getHistoryFilePath(threadId);
-    let initEvent: NavieThreadInitEvent | undefined;
-    const eventLog: TimestampNavieEvent[] = [];
-
-    try {
-      const jsonLines = await readFile(historyFilePath, 'utf-8').then((data) => data.split('\n'));
-      for (const json of jsonLines) {
-        if (json.length === 0) continue;
-        try {
-          const event = JSON.parse(json) as TimestampNavieEvent;
-          if (!initEvent && event.type === 'thread-init') {
-            initEvent = event;
-          }
-          eventLog.push(event);
-        } catch (e) {
-          console.error('Failed to parse event', json, e);
-        }
-      }
-    } catch (e) {
-      throw new Error(`Failed to load history file ${historyFilePath}: ${String(e)}`);
-    }
-
-    if (!initEvent) throw new Error('Thread init event not found');
-
-    const thread = new Thread(initEvent.conversationThread);
-    thread.log = eventLog;
-    thread.lastEventWritten = eventLog.length;
-
-    return thread;
   }
 
   static getHistoryFilePath(threadId: string) {
@@ -167,8 +79,16 @@ export class Thread {
 
   on(event: 'event', clientId: string, listener: (event: NavieEvent) => void): this;
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  on(event: string, clientId: string, listener: (...args: any[]) => void): this {
+  /**
+   * Bind to the given event. A client identifier is used to associate a client with bound
+   * listeners. The listening client must call `removeAllListeners` when it disconnects.
+   *
+   * @param event the event to listen for
+   * @param clientId a unique identifier for the client
+   * @param listener the listener to call when the event is emitted
+   * @returns this
+   */
+  on(event: string, clientId: string, listener: EventListener): this {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-call
     this.eventEmitter.on(event, listener);
 
@@ -184,6 +104,11 @@ export class Thread {
     return this;
   }
 
+  /**
+   * Remove all listeners bound to the given client identifier.
+   *
+   * @param clientId the client identifier to remove
+   */
   removeAllListeners(clientId: string) {
     const listeners = this.listeners.get(clientId);
     if (!listeners) return;
@@ -193,6 +118,9 @@ export class Thread {
     }
   }
 
+  /**
+   * Flush the event log to disk. This is called automatically when a new message is finalized.
+   */
   private async flush() {
     if (this.log.length === this.lastEventWritten) return;
 
@@ -217,22 +145,114 @@ export class Thread {
         }
       }
       const title = lastUserMessage?.content.slice(0, 100);
-      ThreadIndex.getInstance().index(this.conversationThread.id, historyFilePath, title);
+      const threadIndexService = container.resolve(ThreadIndexService);
+      threadIndexService.index(this.conversationThread.id, historyFilePath, title);
     } catch (e) {
       console.error('Failed to update thread index', e);
     }
   }
 
+  /**
+   * Pin an item in the thread. This will emit a `pin-item` event. This item won't actually be
+   * propagated to the backend unless sent by the client via `sendMessage`. This can change in the
+   * future.
+   *
+   * @param item the item to pin
+   */
   pinItem(item: PinnedItem) {
     this.logEvent({ type: 'pin-item', ...item });
   }
 
+  /**
+   * Unpin an item in the thread. This will emit a `unpin-item` event.
+   *
+   * @param item the item to unpin
+   */
   unpinItem(item: PinnedItem) {
     this.logEvent({ type: 'unpin-item', ...item });
   }
 
-  sendMessage(message: string, codeSelection?: UserContext.Context) {
-    const [navie, contextEvents] = NavieService.getNavie();
+  /**
+   * This function is responsible for pre-processing and token emission. It keeps track of active
+   * code blocks and strips out <!-- file: ... --> comments, instead emitting them as metadata.
+   *
+   * @param token The token to be processed and emitted
+   * @param messageId The message id associated with the token
+   */
+  private onToken(token: string, messageId: string) {
+    const subTokens = token.split(/^(`{3,})\n?/gm);
+    for (const subToken of subTokens) {
+      if (subToken.length === 0) continue;
+
+      const fileMatch = subToken.match(/^<!-- file: (.*) -->/);
+      if (fileMatch) {
+        this.codeBlockId = this.codeBlockId ?? randomUUID();
+        this.logEvent({
+          type: 'token-metadata',
+          codeBlockId: this.codeBlockId,
+          metadata: {
+            location: fileMatch[1],
+          },
+        });
+        // Don't emit this token
+        continue;
+      }
+      const language = this.lastTokenBeganCodeBlock ? subToken.match(/^[^\s]+\n/) : null;
+      if (language && this.codeBlockId) {
+        this.logEvent({
+          type: 'token-metadata',
+          codeBlockId: this.codeBlockId,
+          metadata: {
+            language: language[0].trim(),
+          },
+        });
+      }
+
+      this.lastTokenBeganCodeBlock = false;
+
+      let clearCodeBlock = false;
+      if (subToken.match(/^`{3,}/)) {
+        // Code block fences
+        if (this.codeBlockLength === undefined) {
+          this.codeBlockId = this.codeBlockId ?? randomUUID();
+          this.codeBlockLength = subToken.length;
+          this.lastTokenBeganCodeBlock = true;
+        } else if (subToken.length === this.codeBlockLength) {
+          clearCodeBlock = true;
+        }
+      }
+      this.logEvent({
+        type: 'token',
+        token: subToken,
+        messageId,
+        codeBlockId: this.codeBlockId,
+      });
+      if (clearCodeBlock) {
+        this.codeBlockId = undefined;
+        this.codeBlockLength = undefined;
+      }
+    }
+  }
+
+  /**
+   * Send a user message to the thread. This will emit a `message` event. This promise will resolve
+   * once the message has been acknowledged by the backend, before the message is completed.
+   *
+   * @param message the message to send
+   * @param codeSelection (optional) the code selection to send
+   */
+  async sendMessage(
+    message: string,
+    userContext?: NavieRpc.V1.UserContext.ContextItem[]
+  ): Promise<void> {
+    const [navie, contextEvents] = this.navieService.getNavie();
+
+    let context = convertContext(userContext);
+    const { applied, userContext: newUserContext } = await handleReview(message, context);
+    if (applied) {
+      context = newUserContext;
+    }
+
     let responseId: string | undefined;
     contextEvents.on('context', (event) => {
       event.complete
@@ -257,59 +277,7 @@ export class Thread {
         })
         .on('token', (token: string, messageId: string) => {
           if (!responseId) responseId = messageId;
-
-          const subTokens = token.split(/^(`{3,})\n?/gm);
-          for (const subToken of subTokens) {
-            if (subToken.length === 0) continue;
-
-            const fileMatch = subToken.match(/^<!-- file: (.*) -->/);
-            if (fileMatch) {
-              this.codeBlockId = this.codeBlockId ?? randomUUID();
-              this.logEvent({
-                type: 'token-metadata',
-                codeBlockId: this.codeBlockId,
-                metadata: {
-                  location: fileMatch[1],
-                },
-              });
-              // Don't emit this token
-              continue;
-            }
-            const language = this.lastTokenBeganCodeBlock ? subToken.match(/^[^\s]+\n/) : null;
-            if (language && this.codeBlockId) {
-              this.logEvent({
-                type: 'token-metadata',
-                codeBlockId: this.codeBlockId,
-                metadata: {
-                  language: language[0].trim(),
-                },
-              });
-            }
-
-            this.lastTokenBeganCodeBlock = false;
-
-            let clearCodeBlock = false;
-            if (subToken.match(/^`{3,}/)) {
-              // Code block fences
-              if (this.codeBlockLength === undefined) {
-                this.codeBlockId = this.codeBlockId ?? randomUUID();
-                this.codeBlockLength = subToken.length;
-                this.lastTokenBeganCodeBlock = true;
-              } else if (subToken.length === this.codeBlockLength) {
-                clearCodeBlock = true;
-              }
-            }
-            this.logEvent({
-              type: 'token',
-              token: subToken,
-              messageId,
-              codeBlockId: this.codeBlockId,
-            });
-            if (clearCodeBlock) {
-              this.codeBlockId = undefined;
-              this.codeBlockLength = undefined;
-            }
-          }
+          this.onToken(token, messageId);
         })
         .on('error', (err: Error) => {
           this.logEvent({ type: 'error', error: err });
@@ -323,15 +291,27 @@ export class Thread {
             .then(() => this.flush())
             .catch(console.error);
         })
-        .ask(this.conversationThread.id, message, codeSelection, undefined)
+        .ask(this.conversationThread.id, message, context, undefined)
         .catch(reject);
     });
   }
 
+  /**
+   * Gets the events since the given nonce. If no nonce is given, it will return all events.
+   *
+   * @param sinceNonce The nonce to start from
+   * @returns The events since the given nonce
+   */
   getEvents(sinceNonce?: number): readonly TimestampNavieEvent[] {
     return this.log.slice(sinceNonce ?? 0);
   }
 
+  /**
+   * Converts the event log into a chat history. Keep in mind that this processes every event in the
+   * log, so the return value should be cached when possible.
+   *
+   * @returns The chat history
+   */
   getChatHistory(): Navie.ChatHistory {
     const chatHistory: Navie.ChatHistory = [];
     const streamingMessages = new Map<string, Navie.Message>();
@@ -358,171 +338,3 @@ export class Thread {
     return chatHistory;
   }
 }
-
-const threads = new Map<string, Thread>();
-export function registerThread(conversationThread: ConversationThread) {
-  const thread = new Thread(conversationThread);
-  thread.initialize();
-  console.log(`Registered thread ${conversationThread.id}`);
-  threads.set(conversationThread.id, thread);
-}
-
-/**
- * Returns a thread from memory or loads it from disk if it's not yet in memory. If the thread is
- * not found, or the load fails, an error will be thrown.
- * @param threadId the thread identifier to retrieve
- * @returns the thread
- */
-export async function getThread(threadId: string): Promise<Thread> {
-  let thread = threads.get(threadId);
-  if (!thread) {
-    try {
-      thread = await Thread.load(threadId);
-    } catch (e) {
-      // misbehaving threads will be deleted from the index
-      // they probably no longer exist
-      ThreadIndex.getInstance().delete(threadId);
-      throw e;
-    }
-    threads.set(thread.conversationThread.id, thread);
-  }
-  return thread;
-}
-
-const INITIALIZE_SQL = `CREATE TABLE IF NOT EXISTS threads (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    uuid TEXT NOT NULL UNIQUE,
-    path TEXT NOT NULL,
-    title TEXT,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    CONSTRAINT uuid_format CHECK (length(uuid) = 36)
-);
-
-CREATE INDEX IF NOT EXISTS idx_created_at ON threads (created_at);
-CREATE INDEX IF NOT EXISTS idx_uuid ON threads (uuid);
-
-CREATE TABLE IF NOT EXISTS project_directories (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    path TEXT NOT NULL,
-    thread_id TEXT NOT NULL,
-    FOREIGN KEY(thread_id) REFERENCES threads(uuid),
-    UNIQUE (thread_id, path)
-);
-
-CREATE INDEX IF NOT EXISTS idx_thread_id ON project_directories (thread_id);
-`;
-
-const QUERY_INSERT_THREAD_SQL = `INSERT INTO threads (uuid, path, title) VALUES (?, ?, ?)
-ON CONFLICT (uuid) DO UPDATE SET updated_at = CURRENT_TIMESTAMP, title = ?`;
-const QUERY_DELETE_THREAD_SQL = `DELETE FROM threads WHERE uuid = ?`;
-const QUERY_INSERT_PROJECT_DIRECTORY_SQL = `INSERT INTO project_directories (thread_id, path) VALUES (?, ?)
-ON CONFLICT (thread_id, path) DO NOTHING`;
-
-interface QueryOptions {
-  uuid?: string;
-  maxCreatedAt?: Date;
-  orderBy?: 'created_at' | 'updated_at';
-  limit?: number;
-  offset?: number;
-  projectDirectories?: string[];
-}
-
-interface ThreadIndexItem {
-  id: string;
-  path: string;
-  title: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
-
-export class ThreadIndex {
-  private readonly db: sqlite3.Database;
-  private queryInsert: sqlite3.Statement;
-  private queryDelete: sqlite3.Statement;
-  private queryInsertProjectDirectory: sqlite3.Statement;
-
-  private static readonly DATABASE_PATH = join(homedir(), '.appmap', 'navie', 'thread-index.db');
-  private static instance: ThreadIndex;
-
-  private constructor() {
-    this.db = new sqlite3(ThreadIndex.DATABASE_PATH);
-    this.db.exec(INITIALIZE_SQL);
-
-    this.queryInsert = this.db.prepare(QUERY_INSERT_THREAD_SQL);
-    this.queryDelete = this.db.prepare(QUERY_DELETE_THREAD_SQL);
-    this.queryInsertProjectDirectory = this.db.prepare(QUERY_INSERT_PROJECT_DIRECTORY_SQL);
-  }
-
-  static getInstance() {
-    if (!ThreadIndex.instance) {
-      ThreadIndex.instance = new ThreadIndex();
-    }
-    return ThreadIndex.instance;
-  }
-
-  index(threadId: string, path: string, title?: string) {
-    const { projectDirectories } = configuration();
-    this.db.transaction(() => {
-      this.queryInsert.run(threadId, path, title, title);
-
-      // Project directories are written on every index update.
-      //
-      // This is likely not necessary but it'll handle the edge case where an additional project
-      // directory is added mid-way through a conversation.
-      projectDirectories.forEach((projectDirectory) => {
-        this.queryInsertProjectDirectory.run(threadId, projectDirectory);
-      });
-    })();
-    return;
-  }
-
-  delete(threadId: string) {
-    return this.queryDelete.run(threadId);
-  }
-
-  query(options: QueryOptions): ThreadIndexItem[] {
-    let queryString = `SELECT uuid as id, threads.path, title, created_at, updated_at FROM threads`;
-    const params: unknown[] = [];
-    if (options.uuid) {
-      queryString += ` WHERE uuid = ?`;
-      params.push(options.uuid);
-    }
-    if (options.maxCreatedAt) {
-      queryString += ` AND created_at < ?`;
-      params.push(options.maxCreatedAt);
-    }
-    if (options.orderBy) {
-      queryString += ` ORDER BY ? DESC`;
-      params.push(options.orderBy);
-    }
-    if (options.limit) {
-      queryString += ` LIMIT ?`;
-      params.push(options.limit);
-    }
-    if (options.offset) {
-      queryString += ` OFFSET ?`;
-      params.push(options.offset);
-    }
-    if (options.projectDirectories) {
-      if (options.projectDirectories.length === 0) {
-        // If `projectDirectories` is an empty array, we want to return all threads that have no
-        // project directories associated with them. This is edge-casey, but this does occur in
-        // development.
-        //
-        // If you're looking to query for threads with any project directory, leave `projectDirectories`
-        // as `undefined`.
-        queryString += ` LEFT JOIN project_directories ON uuid = thread_id WHERE project_directories.path IS NULL`;
-      } else {
-        queryString += ` INNER JOIN project_directories ON uuid = thread_id WHERE project_directories.path IN (${options.projectDirectories
-          .map(() => '?')
-          .join(',')})`;
-        params.push(...options.projectDirectories);
-      }
-    }
-    const query = this.db.prepare(queryString);
-    return query.all(...params) as ThreadIndexItem[];
-  }
-}
-
-ThreadIndex.getInstance();
