+++ packages/cli/tsconfig.json
@@ -2,6 +2,8 @@
   "compilerOptions": {
     "allowJs": true,
     "esModuleInterop": true,
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
     "lib": ["ES2019", "DOM"],
     "module": "commonjs",
     "moduleResolution": "node",
