+++ packages/components/src/lib/pinnedItems.ts
@@ -1,3 +1,4 @@
+import { NavieRpc } from '@appland/rpc';
 import Vue from 'vue';
 
 interface ObservableContent {
